Added Cassandra support
Now explicitly check that the port is open and available on the provided host. Retry if it is not

Now explicitly check that the port is open and available on the provided host. Retry if it is not

Updated logging

Inverted the control of the Cloud FOundry connection properties for the project

Added MongoDB support